Refactor Schedule view (first part)
