+++ Weekler/Scenes/Schedule/ScheduleView/ScheduleViewController/ScheduleViewController.swift
@@ -15,9 +15,7 @@ import RxCocoa
 final class ScheduleViewController: UIViewController {
     
     //MARK: - private properties
-    private let calendarView: CalendarView
-    private let selectTaskModeView: SelectTaskModeView
-    private let scheduleMainView: ScheduleMainView
+//    private let scheduleMainView: ScheduleMainView!
     private lazy var calendarSwitchRightBarButtonItem: UIBarButtonItem = {
         let button = UIBarButtonItem()
         let image = UIImage(systemName: "calendar")?.withRenderingMode(.alwaysTemplate)
@@ -37,12 +35,6 @@ final class ScheduleViewController: UIViewController {
         self.viewModel = viewModel
         
         //FIXME: create DI method in container
-        let calendarViewModel = viewModel as! CalendarViewModelProtocol
-        let selectTaskViewModel = viewModel as! SelectTaskViewModelProtocol
-        let scheduleMainViewModel = viewModel as! ScheduleMainViewModelProtocol
-        calendarView = CalendarView(frame: .zero, viewModel: calendarViewModel)
-        scheduleMainView = ScheduleMainView(frame: .zero, viewModel: scheduleMainViewModel)
-        selectTaskModeView = SelectTaskModeView(frame: .zero, viewModel: selectTaskViewModel)
         super.init(nibName: nil, bundle: nil)
     }
     
@@ -50,12 +42,18 @@ final class ScheduleViewController: UIViewController {
         fatalError("init(coder:) has not been implemented")
     }
     
+    override func loadView() {
+        super.loadView()
+        let scheduleMainViewModel = viewModel as! ScheduleMainViewModelProtocol
+        self.view = ScheduleMainView(frame: .zero, viewModel: scheduleMainViewModel)
+    }
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         setupUI()
         bind()
-        calendarSwitchRightBarButtonItem.target = calendarView
-        calendarSwitchRightBarButtonItem.action = #selector(CalendarView.calendarSwitchRightBarButtonItemTapped)
+        calendarSwitchRightBarButtonItem.target = self
+        calendarSwitchRightBarButtonItem.action = #selector(calendarSwitchRightBarButtonItemTapped)
     }
     
     override func viewDidLayoutSubviews() {
@@ -71,11 +69,7 @@ final class ScheduleViewController: UIViewController {
     
     //MARK: - private methods
     private func setupUI() {
-//        self.view.backgroundColor = Colors.viewBackground
-        self.view.backgroundColor = .clear
         configureNavBar()
-        addSubviews()
-        applyConstraints()
     }
     
     private func bind() {
@@ -85,7 +79,7 @@ final class ScheduleViewController: UIViewController {
             .subscribe(onNext: { [weak self] height in
                 guard let self = self,
                       let height = height else { return }
-                self.remakeCalendarConstraints(with: height)
+//                self.remakeCalendarConstraints(with: height)
                 self.view.layoutIfNeeded()
             })
             .disposed(by: bag)
@@ -98,50 +92,6 @@ final class ScheduleViewController: UIViewController {
             .disposed(by: bag)
     }
     
-    private func addSubviews() {
-        let subviews = [
-            calendarView,
-            selectTaskModeView,
-            scheduleMainView
-        ]
-        subviews.forEach { view.addSubview($0) }
-    }
-    
-    private func applyConstraints() {
-        //calendarView constraints
-        calendarView.snp.makeConstraints {
-            $0.top.equalTo(view.safeAreaLayoutGuide.snp.top)
-            $0.leading.equalTo(view.safeAreaLayoutGuide.snp.leading)
-            $0.trailing.equalTo(view.safeAreaLayoutGuide.snp.trailing)
-//            $0.height.equalTo(70)
-        }
-        
-        //selectTaskModeView constraints
-        selectTaskModeView.snp.makeConstraints {
-            $0.top.equalTo(calendarView.snp.bottom)
-            $0.leading.equalTo(view.safeAreaLayoutGuide.snp.leading).inset(16)
-            $0.trailing.equalTo(view.safeAreaLayoutGuide.snp.trailing).inset(16)
-            $0.height.equalTo(25)
-        }
-        
-        // scheduleMainView constraints
-        scheduleMainView.snp.makeConstraints {
-            $0.top.equalTo(selectTaskModeView.snp.bottom)
-            $0.leading.equalTo(view.safeAreaLayoutGuide.snp.leading)
-            $0.trailing.equalTo(view.safeAreaLayoutGuide.snp.trailing)
-            $0.bottom.equalTo(view.safeAreaLayoutGuide.snp.bottom)
-        }
-    }
-    
-    private func remakeCalendarConstraints(with height: CGFloat) {
-        calendarView.snp.remakeConstraints {
-            $0.top.equalTo(self.view.safeAreaLayoutGuide.snp.top)
-            $0.leading.equalTo(self.view.safeAreaLayoutGuide.snp.leading)
-            $0.trailing.equalTo(self.view.safeAreaLayoutGuide.snp.trailing)
-            $0.height.equalTo(height)
-        }
-    }
-    
     private func configureNavBar() {
         navigationItem.rightBarButtonItem = calendarSwitchRightBarButtonItem
         if let navBar = navigationController?.navigationBar {
@@ -152,5 +102,11 @@ final class ScheduleViewController: UIViewController {
                 NSAttributedString.Key.foregroundColor: Colors.textColorMain
             ]
         }
+    }
+    
+    @objc
+    private func calendarSwitchRightBarButtonItemTapped() {
+        print("switch")
+        // TODO: - add viewModel.calendarSwitchTapped
     }
 }
